+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Apps - icon in modal should not be red
 - CSS - Force blue color for all huge or massive icons within cards
 - Routing - decoding redirect query param on sign-in page, fix #1133
+- Edges overview - Saves current view as query param, #1125
 
 ## [2.33.1] - 2023-01-30